Branch site options to multiple components according to the chosen flow
