+++ client/landing/stepper/declarative-flow/internals/steps-repository/site-options/index.tsx
@@ -1,292 +1,20 @@
-/* eslint-disable no-nested-ternary */
-/* eslint-disable no-console */
-import { Button, FormInputValidation } from '@automattic/components';
-import { isHostingSiteCreationFlow, StepContainer } from '@automattic/onboarding';
-import { useQuery } from '@tanstack/react-query';
-import { useSelect, useDispatch } from '@wordpress/data';
-import { Icon } from '@wordpress/icons';
-import { useTranslate } from 'i18n-calypso';
-import React, { useEffect } from 'react';
-import siteOptionsUrl from 'calypso/assets/images/onboarding/site-options.svg';
-import storeImageUrl from 'calypso/assets/images/onboarding/store-onboarding.svg';
-import DataCenterPicker from 'calypso/blocks/data-center-picker';
-import DocumentHead from 'calypso/components/data/document-head';
-import FormattedHeader from 'calypso/components/formatted-header';
-import FormFieldset from 'calypso/components/forms/form-fieldset';
-import FormLabel from 'calypso/components/forms/form-label';
-import FormSettingExplanation from 'calypso/components/forms/form-setting-explanation';
-import FormInput from 'calypso/components/forms/form-text-input';
-import FormTextarea from 'calypso/components/forms/form-textarea';
-import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
-import wpcom from 'calypso/lib/wp';
-import { tip } from 'calypso/signup/icons';
-import { useSite } from '../../../../hooks/use-site';
-import { ONBOARD_STORE, SITE_STORE } from '../../../../stores';
+import { VIDEOPRESS_FLOW, isHostingSiteCreationFlow } from '@automattic/onboarding';
+import { NewHostedSiteOptions } from './new-hosted-site-options';
+import { SiteOptions } from './site-options';
+import { VideoPressSiteOptions } from './videopress-site-options';
 import type { Step } from '../../types';
-import type { OnboardSelect } from '@automattic/data-stores';
 import './style.scss';
 
-const getSiteSuggestion = (): Promise< { title: string } > =>
-	wpcom.req.get( '/me/sites/suggestion' );
+const SiteOptionsStepRouter: Step = function SiteOptionsStepRouter( { navigation, flow } ) {
+	if ( isHostingSiteCreationFlow( flow ) ) {
+		return <NewHostedSiteOptions navigation={ navigation } />;
+	}
 
-const SiteOptions: Step = function SiteOptions( { navigation, flow } ) {
-	const { currentSiteTitle, currentTagling } = useSelect(
-		( select ) => ( {
-			currentSiteTitle: ( select( ONBOARD_STORE ) as OnboardSelect ).getSelectedSiteTitle(),
-			currentTagling: ( select( ONBOARD_STORE ) as OnboardSelect ).getSelectedSiteDescription(),
-		} ),
-		[]
-	);
+	if ( flow === VIDEOPRESS_FLOW ) {
+		return <VideoPressSiteOptions navigation={ navigation } />;
+	}
 
-	const { goBack, goNext, submit } = navigation;
-	const [ siteTitle, setSiteTitle ] = React.useState( currentSiteTitle ?? '' );
-	const [ siteGeoAffinity, setSiteGeoAffinity ] = React.useState( '' );
-	const [ tagline, setTagline ] = React.useState( currentTagling ?? '' );
-	const [ formTouched, setFormTouched ] = React.useState( false );
-	const intent = useSelect(
-		( select ) => ( select( ONBOARD_STORE ) as OnboardSelect ).getIntent(),
-		[]
-	);
-	const translate = useTranslate();
-	const site = useSite();
-	const isHostingFlow = isHostingSiteCreationFlow( flow );
-	const isVideoPressFlow = 'videopress' === flow;
-
-	useQuery( [ 'me', 'sites', 'suggestion' ], getSiteSuggestion, {
-		enabled: ! site?.title,
-		onSuccess: ( { title } ) => {
-			if ( ! siteTitle ) {
-				setSiteTitle( title );
-			}
-		},
-	} );
-
-	const { saveSiteSettings } = useDispatch( SITE_STORE );
-
-	useEffect( () => {
-		if ( ! site ) {
-			return;
-		}
-
-		if ( formTouched ) {
-			return;
-		}
-
-		setSiteTitle( site.name || '' );
-		setTagline( site.description );
-	}, [ site, formTouched ] );
-
-	const handleSubmit = async ( event: React.FormEvent ) => {
-		event.preventDefault();
-		if ( site ) {
-			await saveSiteSettings( site.ID, {
-				blogname: siteTitle,
-				blogdescription: tagline,
-			} );
-		}
-
-		if ( isVideoPressFlow || site || isHostingFlow ) {
-			recordTracksEvent( 'calypso_signup_site_options_submit', {
-				has_site_title: !! siteTitle,
-				has_tagline: !! tagline,
-			} );
-
-			submit?.( { siteTitle, tagline, siteGeoAffinity } );
-		}
-	};
-
-	const onChange = ( event: React.FormEvent< HTMLInputElement > ) => {
-		if ( isVideoPressFlow || site || isHostingFlow ) {
-			setFormTouched( true );
-
-			switch ( event.currentTarget.name ) {
-				case 'siteTitle':
-					return setSiteTitle( event.currentTarget.value );
-				case 'tagline':
-					return setTagline( event.currentTarget.value );
-			}
-		}
-	};
-
-	const getTextsForVideoPressFlow = () => {
-		return {
-			headerText: translate( 'Set up your video site' ),
-			headerImage: undefined,
-			siteTitleLabel: translate( 'Site name' ),
-			taglineLabel: translate( 'Brief description' ),
-			taglineExplanation: translate( 'Add a short description of your video site here.' ),
-			subHeaderText: translate( 'Customize some details about your new site.' ),
-		};
-	};
-
-	const getTextsForHostingFlow = () => {
-		return {
-			headerText: translate( "Let's create your site" ),
-			headerImage: siteOptionsUrl,
-			siteTitleLabel: translate( 'Site title' ),
-			taglineLabel: undefined,
-			taglineExplanation: undefined,
-			subHeaderText: undefined,
-		};
-	};
-
-	const getTextsFromIntent = ( intent: string ) => {
-		switch ( intent ) {
-			case 'sell':
-				return {
-					headerText: translate( "First, let's give your store a name" ),
-					headerImage: storeImageUrl,
-					siteTitleLabel: translate( 'Store name' ),
-					taglineLabel: translate( 'Tagline' ),
-					taglineExplanation: translate( 'In a few words, explain what your store is about.' ),
-					subHeaderText: undefined,
-				};
-			case 'write':
-			default:
-				return {
-					headerText: translate( "First, let's give your blog a name" ),
-					headerImage: siteOptionsUrl,
-					siteTitleLabel: translate( 'Blog name' ),
-					taglineLabel: translate( 'Tagline' ),
-					taglineExplanation: translate( 'In a few words, explain what your blog is about.' ),
-					subHeaderText: undefined,
-				};
-		}
-	};
-
-	const isSiteTitleRequired = isVideoPressFlow || isHostingFlow;
-	const isTaglineRequired = isVideoPressFlow;
-
-	const taglineError = null;
-
-	const getTextsForFlow = () => {
-		if ( isHostingFlow ) {
-			return getTextsForHostingFlow();
-		}
-
-		return isVideoPressFlow ? getTextsForVideoPressFlow() : getTextsFromIntent( intent );
-	};
-
-	const textsForFlow = getTextsForFlow();
-
-	const {
-		headerText,
-		headerImage,
-		siteTitleLabel,
-		taglineLabel,
-		taglineExplanation,
-		subHeaderText,
-	} = textsForFlow;
-
-	const isFormDisabled = ! isVideoPressFlow && ! site && ! isHostingFlow;
-	const isSiteTitleEmpty = ! siteTitle || siteTitle.trim().length === 0;
-	const isFormSubmitDisabled =
-		isFormDisabled || ( ( isVideoPressFlow || isHostingFlow ) && isSiteTitleEmpty );
-	const siteTitleError = formTouched && isSiteTitleRequired && isSiteTitleEmpty;
-
-	const stepContent = (
-		<form className="site-options__form" onSubmit={ handleSubmit }>
-			<FormFieldset disabled={ isFormDisabled } className="site-options__form-fieldset">
-				<FormLabel htmlFor="siteTitle" optional={ ! isSiteTitleRequired }>
-					{ siteTitleLabel }
-				</FormLabel>
-				<FormInput
-					name="siteTitle"
-					id="siteTitle"
-					value={ siteTitle }
-					isError={ siteTitleError }
-					onChange={ onChange }
-					placeholder={
-						isVideoPressFlow
-							? translate( 'My Video Site' )
-							: isHostingFlow
-							? translate( 'My Hosted Site' )
-							: null
-					}
-				/>
-				{ siteTitleError ? (
-					<FormInputValidation isError text={ translate( 'Please provide a site title' ) } />
-				) : isHostingFlow ? (
-					<FormSettingExplanation className="site-title-optional-explanation">
-						<Icon className="site-title-optional-explanation__icon" icon={ tip } size={ 24 } />
-						{ translate( "Don't worry, you can change it later." ) }
-					</FormSettingExplanation>
-				) : null }
-			</FormFieldset>
-			{ isHostingFlow && (
-				<DataCenterPicker onChange={ setSiteGeoAffinity } value={ siteGeoAffinity } compact />
-			) }
-			{ taglineLabel && (
-				<FormFieldset disabled={ isFormDisabled } className="site-options__form-fieldset">
-					<FormLabel htmlFor="tagline" optional={ ! isTaglineRequired }>
-						{ taglineLabel }
-					</FormLabel>
-					{ ! isVideoPressFlow && (
-						<FormInput
-							name="tagline"
-							id="tagline"
-							value={ tagline }
-							isError={ taglineError }
-							onChange={ onChange }
-							placeholder={ null }
-						/>
-					) }
-					{ isVideoPressFlow && (
-						<FormTextarea
-							name="tagline"
-							id="tagline"
-							value={ tagline }
-							isError={ taglineError ?? undefined }
-							onChange={ onChange }
-							placeholder={ taglineExplanation }
-						/>
-					) }
-					{ taglineError && <FormInputValidation isError text={ taglineError } /> }
-					{ ! isVideoPressFlow && (
-						<FormSettingExplanation>
-							<Icon className="site-options__form-icon" icon={ tip } size={ 20 } />
-							{ taglineExplanation }
-						</FormSettingExplanation>
-					) }
-				</FormFieldset>
-			) }
-			<Button
-				disabled={ isFormSubmitDisabled }
-				className="site-options__submit-button"
-				type="submit"
-				primary
-			>
-				{ translate( 'Continue' ) }
-			</Button>
-		</form>
-	);
-
-	return (
-		<>
-			<DocumentHead title={ headerText } />
-			<StepContainer
-				stepName="site-options"
-				shouldHideNavButtons={ isVideoPressFlow || isHostingFlow }
-				className={ `is-step-${ intent }` }
-				headerImageUrl={ headerImage }
-				hideSkip={ true }
-				goBack={ goBack }
-				goNext={ goNext }
-				isHorizontalLayout={ ! isVideoPressFlow }
-				formattedHeader={
-					<FormattedHeader
-						id="site-options-header"
-						headerText={ headerText }
-						align="left"
-						subHeaderText={ subHeaderText }
-					/>
-				}
-				stepContent={ stepContent }
-				recordTracksEvent={ recordTracksEvent }
-				showVideoPressPowered={ isVideoPressFlow }
-			/>
-		</>
-	);
+	return <SiteOptions navigation={ navigation } />;
 };
 
-export default SiteOptions;
+export default SiteOptionsStepRouter;
